// projects/ui-lib-custom/src/lib/dialog/dialog.types.ts
/**
 * Supported dialog positions.
 */
export type DialogPosition =
  | 'center'
  | 'top'
  | 'bottom'
  | 'left'
  | 'right'
  | 'top-left'
  | 'top-right'
  | 'bottom-left'
  | 'bottom-right';

/**
 * Supported dialog variants.
 */
export type DialogVariant = 'material' | 'bootstrap' | 'minimal';